+++ Public/Functions/Catalogs/Dell/Get-DellSystemCatalog.ps1
@@ -86,7 +86,7 @@ function Get-DellSystemCatalog {
     $RawCatalogCabName  	= [string]($OnlineCatalogUri | Split-Path -Leaf)
     $RawCatalogCabPath 		= Join-Path $env:TEMP (Join-Path 'OSD' $RawCatalogCabName)
     $TempCatalogFile        = Join-Path $env:TEMP (Join-Path 'OSD' $OfflineCatalogName)
-    $ModuleCatalogFile      = "$($MyInvocation.MyCommand.Module.ModuleBase)\Catalogs\$OfflineCatalogName"
+    $ModuleCatalogFile      = "$(Get-OSDCatsPath)\osd-module\$OfflineCatalogName"
     #=================================================
     #   UseCatalog Cloud
     #=================================================