+++ Public/Functions/Catalogs/Intel/Get-IntelRadeonDriverPack.ps1
@@ -42,7 +42,7 @@ function Get-IntelRadeonDriverPack {
     }
 
     $TempCatalogFile = Join-Path $env:TEMP (Join-Path 'OSD' $OfflineCatalogName)
-    $ModuleCatalogFile = "$($MyInvocation.MyCommand.Module.ModuleBase)\Catalogs\$OfflineCatalogName"
+    $ModuleCatalogFile = "$(Get-OSDCatsPath)\osd-module\$OfflineCatalogName"
     $ModuleCatalogContent = Get-Content -Path $ModuleCatalogFile -Raw | ConvertFrom-Json
     #=================================================
     #   IsOnline