+++ Public/OSDCloudTS/Get-OSDCloudOperatingSystemsIndexMap.ps1
@@ -6,6 +6,9 @@ function Get-OSDCloudOperatingSystemsIndexMap {
     .DESCRIPTION
     Returns the Operating System Indexes used by OSDCloud
 
+    .PARAMETER OSArch
+    Specifies the OS architecture to filter results. Valid values are 'x64' and 'ARM64'.
+
     .LINK
     https://github.com/OSDeploy/OSD/tree/master/Docs
     #>
@@ -13,18 +16,14 @@ function Get-OSDCloudOperatingSystemsIndexMap {
     [CmdletBinding()]
     param (
         [Parameter(ParameterSetName = 'Default')]
-        [ValidateSet('x64','ARM64')]
+        [ValidateSet('x64', 'ARM64')]
         [System.String]
         $OSArch = 'x64'
-
     )
-    if ($OSArch -eq 'x64'){
-        $Results = Get-Content -Path (Join-Path (Get-Module -Name OSD -ListAvailable | Sort-Object Version -Descending | Select-Object -First 1).ModuleBase "Catalogs\CloudOperatingIndexMap.json") | ConvertFrom-Json
-        $Results = $Results | Where-Object {$_.Architecture -eq 'x64'}
-    }
-    elseif ($OSArch -eq "ARM64"){
-        $Results = Get-Content -Path (Join-Path (Get-Module -Name OSD -ListAvailable | Sort-Object Version -Descending | Select-Object -First 1).ModuleBase "Catalogs\CloudOperatingIndexMap.json") | ConvertFrom-Json
-        $Results = $Results | Where-Object {$_.Architecture -eq 'ARM64'}
-    }
-    $Results
+
+    $indexMapPath = "$(Get-OSDCatsPath)\osd-module\CloudOperatingIndexMap.json"
+    $Results = Get-Content -Path $indexMapPath | ConvertFrom-Json
+    $Results = $Results | Where-Object { $_.Architecture -eq $OSArch }
+    
+    return $Results
 }